Move function body generation to goto instrument directory
This is because we want to use the c nondet symbol factory from the ansi-c
directory, and code in goto-programs should not depend on code in ansi-c.
