+++ src/goto-instrument/generate_function_bodies.h
